schema for doctors exported to separate file, models folder created
